Drop the workspace outline's stack header row
The header named the stack and held a kebab, but its drag glyph was
decorative — stack reordering was never implemented, and its Move Up/Down
items were hard-disabled. The card's own border already separates one
stack from the next, as the branches tab shows.

Its remaining actions move into every branch's menu as their own group:
folding the whole stack, updating it, and unapplying it — the last two
had no other path at all.
